LayoutTests/imported/w3c/web-platform-tests/xhr/responsexml-document-properties-expected.txt


FAIL domain assert_equals: expected (undefined) undefined but got (string) "localhost"
PASS URL 
PASS documentURI 
PASS baseURI 
PASS referrer 
PASS title 
PASS contentType 
FAIL readyState assert_equals: expected "complete" but got "interactive"
PASS location 
PASS defaultView 
FAIL body assert_equals: expected (undefined) undefined but got (object) null
FAIL images assert_equals: expected (undefined) undefined but got (object) object "[object HTMLCollection]"
PASS doctype 
FAIL forms assert_equals: expected (undefined) undefined but got (object) object "[object HTMLCollection]"
FAIL all assert_equals: expected undefined but got [object HTMLAllCollection]
FAIL links assert_equals: expected (undefined) undefined but got (object) object "[object HTMLCollection]"
PASS cookie 
FAIL Test document URL properties after redirect assert_equals: expected "http://localhost:8800/XMLHttpRequest/resources/well-formed.xml" but got "http://localhost:8800/XMLHttpRequest/resources/redirect.py?location=well-formed.xml"
FAIL Test document URL properties of document with <base> after redirect assert_equals: expected "http://localhost:8800/XMLHttpRequest/resources/base.xml" but got "http://localhost:8800/XMLHttpRequest/resources/redirect.py?location=base.xml"
PASS lastModified set to time of response if no HTTP header provided 
FAIL lastModified set to related HTTP header if provided assert_equals: expected 1505770096000 but got 1505744052000
PASS cookie (after setting it) 
PASS styleSheets 
PASS implementation 
